Added checkRequirements and resets to EightBitter
Each EightBitter may have a set of reset functions (called in order in
the reset constructor) and requirements (global, checked for in the
window; self, checked for under "this") that they check for in the reset
constructor. FullScreenMario has all the modules under /src as globals,
and sprites.js under self.