Fleshed out MapsCreatr, MapsHandlr
The two are close to being ready to take over the legacy MapsManagr.

map_settings references have been removed from FullScreenMario.js -
MapsHandlr correctly copies the required information to the MapScreener.

As a small side note, setMap also calls FSM.MapsHandler.setMap("1-1") -
this is so it copies the info to MapScreener when needed.