Solidified build process
Added /dist to gitignore because it now has copies of Sounds, Theme, and
Fonts. In the future it would be nice to replace Fonts with PixelRendr &
PixelDrawer and Sounds with a JSON-based AudioPlayr (see commits related
to AudioPlayr/AudioPlayrNew).

Moved ImageReadr into /src/utils because that's where it belongs.

Removed old, unused images from Theme.

The root index.html now works and has Grunt processing it into the
dist/index.html.

Fixes #56.